add swipe to delete for sets
